CON-2814: Refactored the PaymentBar to enable dynamic CartPage padding - introduced react-sizeme module for component size measuring - refactored the PaymentBarComponent to make it usable with react-sizeme - updated the CartPageComponent and it's style generation to calculate a dynamic bottom padding with the current size of the PaymentBar
