+++ pages/Cart/components/PaymentBar/components/Content/components/SubTotal/connector.js
