+++ pages/Cart/components/PaymentBar/components/Content/components/SubTotal/style.js
