+++ pages/Cart/components/PaymentBar/components/Content/components/SubTotalLabel/index.jsx
