+++ pages/Cart/components/PaymentBar/components/Content/components/SubTotalLabel/style.js
