+++ pages/Cart/components/PaymentBar/spec.jsx
@@ -8,14 +8,15 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import I18n from '@shopgate/pwa-common/components/I18n';
-import SubTotal from './components/SubTotal';
-import ShippingCosts from './components/ShippingCosts';
+import SubTotal from './components/Content/components/SubTotal';
+import ShippingCosts from './components/Content/components/ShippingCosts';
+import Content from './components/Content/index';
 import PaymentBar from './index';
 
 /**
  * Mock the connect() methods of the relevant sub-components.
  */
-jest.mock('./components/CheckoutButton/connector', () => (obj) => {
+jest.mock('./components/Content/components/CheckoutButton/connector', () => (obj) => {
   const newObj = obj;
 
   newObj.defaultProps = {
@@ -27,7 +28,7 @@ jest.mock('./components/CheckoutButton/connector', () => (obj) => {
   return newObj;
 });
 
-jest.mock('./components/ShippingCosts/connector', () => (obj) => {
+jest.mock('./components/Content/components/ShippingCosts/connector', () => (obj) => {
   const newObj = obj;
 
   newObj.defaultProps = {
@@ -40,7 +41,7 @@ jest.mock('./components/ShippingCosts/connector', () => (obj) => {
   return newObj;
 });
 
-jest.mock('./components/ShippingCostsLabel/connector', () => (obj) => {
+jest.mock('./components/Content/components/ShippingCostsLabel/connector', () => (obj) => {
   const newObj = obj;
 
   newObj.defaultProps = {
@@ -52,7 +53,7 @@ jest.mock('./components/ShippingCostsLabel/connector', () => (obj) => {
   return obj;
 });
 
-jest.mock('./components/SubTotal/connector', () => (obj) => {
+jest.mock('./components/Content/components/SubTotal/connector', () => (obj) => {
   const newObj = obj;
 
   newObj.defaultProps = {
@@ -65,7 +66,7 @@ jest.mock('./components/SubTotal/connector', () => (obj) => {
   return newObj;
 });
 
-jest.mock('./components/SubTotalLabel/connector', () => (obj) => {
+jest.mock('./components/Content/components/SubTotalLabel/connector', () => (obj) => {
   const newObj = obj;
 
   newObj.defaultProps = {
@@ -88,11 +89,10 @@ describe('<PaymentBar />', () => {
    * @param {Object} props The component props.
    * @return {Object} The mounted component.
    */
-  const renderComponent = (props = {}) => mount(
+  const renderComponent = (props = {}) => mount((
     <I18n.Provider lang={langCode} locales={testLocales}>
       <PaymentBar {...props} currency="USD" />
-    </I18n.Provider>
-  );
+    </I18n.Provider>));
 
   it('should render without any props', () => {
     const wrapper = renderComponent();
@@ -101,7 +101,7 @@ describe('<PaymentBar />', () => {
 
   it('should show the formatted subtotal', () => {
     const wrapper = renderComponent({});
-    const component = wrapper.find(PaymentBar);
+    const component = wrapper.find(Content);
 
     expect(wrapper).toMatchSnapshot();
     expect(component.find(SubTotal).text()).toEqual('$4.99');
@@ -109,7 +109,7 @@ describe('<PaymentBar />', () => {
 
   it('should show formatted shipping costs', () => {
     const wrapper = renderComponent({});
-    const component = wrapper.find(PaymentBar);
+    const component = wrapper.find(Content);
 
     expect(wrapper).toMatchSnapshot();
     expect(component.find(ShippingCosts).text()).toEqual('$2.99');
