+++ pages/Cart/index.jsx
@@ -16,7 +16,6 @@ import Empty from './components/Empty';
 import PaymentBar from './components/PaymentBar';
 import connect from './connector';
 import styles from './style';
-
 /**
  * The cart view component.
  * @returns {JSX}
@@ -46,7 +45,18 @@ class Cart extends Component {
 
     this.state = {
       isPaymentBarHidden: false,
+      containerPaddingStyle: styles.getContainerPaddingStyle(),
     };
+  }
+
+  /**
+   * Callback for the onSize event of the PaymentBar.
+   * @param {Object} size An object which contains data about the current componenent dimensions.
+   */
+  onSize = ({ height }) => {
+    this.setState({
+      containerPaddingStyle: styles.getContainerPaddingStyle(height),
+    });
   }
 
   /**
@@ -78,7 +88,7 @@ class Cart extends Component {
 
     return (
       <View title={this.title}>
-        <section className={styles.container}>
+        <section className={styles.container} style={this.state.containerPaddingStyle}>
           {messages.length > 0 && <MessageBar messages={messages} />}
           {cartItems.length > 0 && (
             <Fragment>
@@ -92,7 +102,7 @@ class Cart extends Component {
                 ))}
                 <CouponField onToggleFocus={this.togglePaymentBar} />
               </CardList>
-              <PaymentBar isVisible={!this.state.isPaymentBarHidden} />
+              <PaymentBar isVisible={!this.state.isPaymentBarHidden} onSize={this.onSize} />
             </Fragment>
           )}
         </section>
